fix(ffi): change alec_encode_multi values param from f64 to f32
The Rust FFI signature declared `values: *const f64` (8 bytes per
element) while the C header declared `const float*` (4 bytes per
element). On Cortex-M33 this caused Rust to read 8 bytes per value
from a buffer of 4-byte floats, overreading allocated memory (UB)
and producing ALEC_ERR_NULL_POINTER (rc=5).

Fix: accept *const f32 and widen to f64 inside the FFI shim before
passing to Encoder::encode_multi().

+++ alec-ffi/include/alec.h
@@ -174,7 +174,7 @@ AlecResult alec_encode_value(
  */
 AlecResult alec_encode_multi(
     AlecEncoder* encoder,
-    const double* values,
+    const float* values,
     size_t value_count,
     uint64_t timestamp,
     const char* source_id,
